README and EXERCISES

+++ modulo2/ex19/asm.h
@@ -1,4 +1,4 @@
 #ifndef ASM_H
 #define ASM_H
-short needed_time(void);
+int needed_time(void);
 #endif
